requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
Werkzeug>=3.0.0
gunicorn>=21.0.0
